Initial revision of Thaana (`Thaa`) script for Dhivehi.
As a side change, introducing protocol buffer describing Brahmic script-specific configuration parameters. This is very helpful for Thaana which, while possessing some features common with Brahmic scripts, nevertheless does not strictly qualify as an abugida. In particular, it does not possess inherent vowels.

### Notes:

1.  In occasional words found online, some of the consonant clusters are missing `SUKUN`, which is currently disallowed by the ISO and well-formedness grammar. I've
checked 50K word list from [An Crúbadán](http://crubadan.org/) and we are getting 1,005 *well-formedness* failures, mostly in loan words and foreign names. Most of these are due to omitted `SUKUN`. On the same word list, the number of romanisation failures is 900.
1.  At the moment, similar to other Brahmic scripts, the vowel hiatus is marked with a `.` in our romanisations of Dhivehi.
1.  Some of the words for test are taken from Divehi wordlist at [An Crúbadán](http://crubadan.org/).

PiperOrigin-RevId: 401246711