Ensures that the random generated words tests work in the native-to-Latin direction for reversible romanization (ISO).
### Telugu

The character `CANDRABINDU` has multiple definitions in both `nisaba/scripts/brahmic/data/Telu/coda.textproto` and `nisaba/scripts/brahmic/data/coda.textproto`. So, it has been removed from `nisaba/scripts/brahmic/data/Telu/coda.textproto`.

### Test size

The test size has been changed from "enormous" to "medium" as per the build system's recommendation.

### Kannada
The weight of the rule for marking explicit virama has been reduced further. This is because the Kannada sequence `<AA, VIRAMA, YA>` (which has virama) acts as a vowel. Therefore, the weight of the rule for independent vowels must be higher than that for the rule handling the explicit virama rule. However, the weight for the independent vowels rule is already low, as it should not conflict with the vowel signs. Therefore, the weight of the explicit virama rule had to be reduced further.

### Meitei

The `SIGN NUNG` was defined multiple times. Therefore, it has been removed from `vowel.textproto` to be handled like Anusvara in the rest of the Brahmic scripts through `coda.textproto`.

PiperOrigin-RevId: 547609938