To make randgen tests faster, `FROM_ARAB` and `TO_ARAB` are no longer composed together in the romanization round-trip test.
Additionally, the `reversible_roman` of `abjad_alphabet` is made analogous to that of Brahmic ISO. This involved removing the ability to accept any string as input. It now expects only Arabic characters, like any Brahmic script ISO.

NFC is not needed for TO_ARAB. However, this is needed for Brahmic for Gurmukhi `SHA` and its NFC form `<SA, NUKTA>` which apparently has the same romanization defined, unlike Devanagari `QA`.

Unlike Brahmic, NFC for Arabic is not idempotent. The current NFC code cannot handle the reordering of a large number of `SHADDA`, `FATHA`, `KASRA`, etc.. Even though, this is only a theoretical possibility, for randgen tests, the number of NFCs we do as part of the round tripping should be the same as the number of NFCs applied on the input before verification.

PiperOrigin-RevId: 550303033